refactor: move new getSegmentInfo and getSegmentVariantInfo to info.js
